added food_log2.dart and inference.dart
